+++ api/internal/handler/remoteapisubscription/syncer/syncer.go
@@ -80,10 +80,10 @@ func (c *syncerClient) Send(ctx context.Context, resource *apiv1.RemoteApiSubscr
 			},
 		},
 	}
-	if resource.Spec.Security != nil {
+	if resource.HasM2M() {
 		body.Security = &cpv1.Security{
 			Oauth2: &cpv1.SecurityOauth2{
-				Scopes: &resource.Spec.Security.Oauth2Scopes,
+				Scopes: &resource.Spec.Security.M2M.Scopes,
 			},
 		}
 	}