+++ gateway/internal/features/builder_test.go
@@ -18,106 +18,10 @@ import (
 	"github.com/telekom/controlplane/gateway/pkg/kong/client/mock"
 	"github.com/telekom/controlplane/gateway/pkg/kong/client/plugin"
 	"go.uber.org/mock/gomock"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func NewMockRoute() *gatewayv1.Route {
-	return &gatewayv1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.RouteSpec{
-			Realm: types.ObjectRef{
-				Name:      "realm",
-				Namespace: "default",
-			},
-			PassThrough: false,
-			Upstreams: []gatewayv1.Upstream{
-				{
-					Scheme: "http",
-					Host:   "upstream.url",
-					Port:   8080,
-					Path:   "/api/v1",
-				},
-			},
-			Downstreams: []gatewayv1.Downstream{
-				{
-					Host:      "downstream.url",
-					Port:      8080,
-					Path:      "/test/v1",
-					IssuerUrl: "issuer.url",
-				},
-			},
-		},
-	}
-}
-
-func NewMockConsumeRoute(routeRef types.ObjectRef) *gatewayv1.ConsumeRoute {
-	return &gatewayv1.ConsumeRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-consumer",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.ConsumeRouteSpec{
-			ConsumerName: "test-consumer-name",
-			Route:        routeRef,
-		},
-	}
-}
-
-func NewMockRealm() *gatewayv1.Realm {
-	return &gatewayv1.Realm{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-realm",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.RealmSpec{
-			Url:       "https://realm.url",
-			IssuerUrl: "https://issuer.url",
-			DefaultConsumers: []string{
-				"gateway",
-				"test",
-			},
-		},
-	}
-}
-
-func NewMockGateway() *gatewayv1.Gateway {
-	return &gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-gateway",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			Admin: gatewayv1.AdminConfig{
-				ClientId:     "admin",
-				ClientSecret: "topsecret",
-				IssuerUrl:    "https://issuer.url",
-				Url:          "https://admin.test.url",
-			},
-		},
-	}
-}
-
 var _ = Describe("FeatureBuilder", Ordered, func() {
-	var mockCtrl *gomock.Controller
-	BeforeAll(func() {
-		mockCtrl = gomock.NewController(GinkgoT())
-	})
-
 	Context("Registering", Ordered, func() {
-
-		var route *gatewayv1.Route
-		var realm *gatewayv1.Realm
-		var gateway *gatewayv1.Gateway
-
-		BeforeAll(func() {
-			route = NewMockRoute()
-			realm = NewMockRealm()
-			gateway = NewMockGateway()
-		})
-
 		It("should be registered", func() {
 			kc := mock.NewMockKongClient(mockCtrl)
 
@@ -126,31 +30,20 @@ var _ = Describe("FeatureBuilder", Ordered, func() {
 			builder.EnableFeature(feature.InstancePassThroughFeature)
 			builder.EnableFeature(feature.InstanceAccessControlFeature)
 			builder.EnableFeature(feature.InstanceLastMileSecurityFeature)
+			builder.EnableFeature(feature.InstanceCustomScopesFeature)
+			builder.EnableFeature(feature.InstanceLoadBalancingFeature)
+			builder.EnableFeature(feature.InstanceExternalIDPFeature)
 
 			b, ok := builder.(*features.Builder)
 			Expect(ok).To(BeTrue())
-			Expect(b.Features).To(HaveLen(3))
+			Expect(b.Features).To(HaveLen(6))
 		})
 
 	})
 
 	Context("Applying and Creating", Ordered, func() {
-
 		var ctx = context.Background()
-		var mockKc *mock.MockKongClient
-
-		var route *gatewayv1.Route
-		var realm *gatewayv1.Realm
-		var gateway *gatewayv1.Gateway
-
-		BeforeAll(func() {
-			route = NewMockRoute()
-			realm = NewMockRealm()
-			gateway = NewMockGateway()
-
-			ctx = contextutil.WithEnv(ctx, "test")
-		})
-
+		ctx = contextutil.WithEnv(ctx, "test")
 		BeforeEach(func() {
 			mockKc = mock.NewMockKongClient(mockCtrl)
 		})
@@ -268,7 +161,7 @@ var _ = Describe("FeatureBuilder", Ordered, func() {
 
 			Expect(rtPlugin.Config.Append.Headers.Get("remote_api_url")).To(Equal("http://upstream.url:8080/api/v1"))
 			Expect(rtPlugin.Config.Append.Headers.Get("api_base_path")).To(Equal("/api/v1"))
-			Expect(rtPlugin.Config.Append.Headers.Get("jumper_config")).To(Equal("e30="))
+			Expect(b.JumperConfig()).To(Equal(plugin.NewJumperConfig())) // empty jumper config
 
 			Expect(rtPlugin.Config.Add.Headers.Get("environment")).To(Equal("test"))
 			Expect(rtPlugin.Config.Add.Headers.Get("realm")).To(Equal("test-realm"))
@@ -315,11 +208,35 @@ var _ = Describe("FeatureBuilder", Ordered, func() {
 
 			By("checking the request-transformer plugin config")
 			Expect(rtPlugin.Config.Append.Headers.Get("remote_api_url")).To(Equal("http://upstream.url:8080/api/v1"))
-			Expect(rtPlugin.Config.Append.Headers.Get("jumper_config")).To(Equal("e30="))
 
 			Expect(rtPlugin.Config.Append.Headers.Get("issuer")).To(Equal("https://upstream.issuer.url"))
 			Expect(rtPlugin.Config.Append.Headers.Get("client_id")).To(Equal("gateway"))
 			Expect(rtPlugin.Config.Append.Headers.Get("client_secret")).To(Equal("topsecret"))
+		})
+
+		It("should apply the CustomScopes feature", func() {
+			scopesRoute := route.DeepCopy()
+			consumeRoute := NewMockConsumeRoute(*types.ObjectRefFromObject(scopesRoute))
+			builder := features.NewFeatureBuilder(mockKc, scopesRoute, realm, gateway)
+			builder.AddAllowedConsumers(consumeRoute)
+			builder.EnableFeature(feature.InstanceCustomScopesFeature)
+			builder.EnableFeature(feature.InstanceLastMileSecurityFeature)
+
+			mockKc.EXPECT().CreateOrReplaceRoute(ctx, scopesRoute, gomock.Any()).Return(nil).Times(1)
+			mockKc.EXPECT().CreateOrReplacePlugin(ctx, gomock.Any()).Return(nil, nil).Times(1)
+			mockKc.EXPECT().CleanupPlugins(ctx, gomock.Any(), gomock.Any()).Return(nil).Times(1)
+
+			By("building the features")
+			err := builder.Build(ctx)
+			Expect(err).ToNot(HaveOccurred())
+
+			b, ok := builder.(*features.Builder)
+			Expect(ok).To(BeTrue())
+
+			By("Checking that jumperConfig is filled with the scopes")
+			Expect(b.JumperConfig).ToNot(BeNil())
+			Expect(b.JumperConfig().OAuth[plugin.ConsumerId("default")].Scopes).To(Equal("scope1"))
+			Expect(b.JumperConfig().OAuth[plugin.ConsumerId(consumeRoute.Spec.ConsumerName)].Scopes).To(Equal("scope1 scope2"))
 		})
 
 		It("should correctly apply the RateLimit feature", func() {