+++ rover/cmd/main.go
@@ -35,6 +35,7 @@ import (
 	applicationv1 "github.com/telekom/controlplane/application/api/v1"
 	organizationv1 "github.com/telekom/controlplane/organization/api/v1"
 	roverv1 "github.com/telekom/controlplane/rover/api/v1"
+
 	//+kubebuilder:scaffold:imports
 
 	secretmetrics "github.com/telekom/controlplane/secret-manager/api/metrics"